🧪 Let `ansible-test` diff ancient branches
This patch addresses the long-standing issue with being unable to run
CI on PRs that have their branches created over 500 commits prior.

When the first `git diff` attempt fails, this patch attempts to
augment the Git tree with missing commits.

This is achieved by doing the following:
1. Fetch one additional commit at the root of each orphaned
   branch.
2. Identify said commits as tree roots or "grafted".
3. Retrieve the timestamps of both and safe the oldest one.
4. Fetch again, now instructing Git to include everything past
   this date.

The idea is that the very first commit in the PR branch might
be created on the same day as the PR is opened. However, its
parent would belong to the base branch, which is immutable. So
if we take that commit's time stamp, we can tell Git to fill
in the commits between the PR branch chunk and the commit it
branched out of (the fork point). We'll also fill in the
commits on the other side of the "fork", in the base branch.
With those commits retrieved and available in the local cache
of the remote, it will be possible for git diff to locate the
merge base and compute the file list, while the tree remains
shallow. This is still a heuristic, of course, since the
commit dates can be out of order in the tree. But it's not
expected that we'll hit this corner case often or ever.